app/requirements.txt
daphne==4.1.2  # ASGI runserver for development
Django==5.1.5
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-simple-plausible==0.0.5
ipython==8.31.0
paho-mqtt==2.1.0
redis==5.2.1
uvicorn[standard]==0.32.1